requirements.txt
pandas
numpy
faker
psycopg2-binary
dbt-postgres
python-dotenv
